requirements.txt
google-play-scraper==1.2.7
google-generativeai==0.8.3
pandas==2.2.3
sentence-transformers==3.2.1
scikit-learn==1.5.2
python-dotenv==1.0.1
tqdm==4.67.0
colorama
